requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.1.17
langchain-community==0.0.38
langchain-google-genai==0.0.6
pinecone-client==4.1.2
python-dotenv==1.0.1
pandas==2.2.2
rank-bm25==0.2.2
mangum==0.17.0
